+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased]
+
+### Added
+- implement `isContinuous`.
+- send `primaryActionTriggered` event in addition to `valueChanged`.
+
 ## [1.5.1] - 2018-07-03
 
 ### Fixed