+++ CHANGELOG.md
@@ -4,7 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.7.0]
+
+### Changed
+- Swift 4.2
+- animate thumb movement to make it smoother.
 
 ### Fixed
 - zero-width sliding area in iOS 12 caused all thumbs to be shown on left and crash on drag.